Improve ZSH version of CLI instructions
Use the 'j' operator to allow users to specify either in the form
`gi python,vim` or `gi python vim`.

On slow networks, some curl installs default to showing progress which
is likely not what the user wants. Use the "-s" option to make curl only
show output for errors.