Update #437 - accept "+" letter in url parameter
Pull request #437 actually didn't solve the closed issue #435 because I understood the problem wrong. This problem was not about url encoding.

`URLSearchParams` object returns invalid gitignore template id not because the object didn't recognize `+` letters, but because it ACTUALLY recognized them and switched them to SPACES(` `).

To solve the problem correctly, this commit replaces spaces to `+` letters so the filter is able to find proper ids from template set.