Fix #437 - the "+" letter problem (#438)
* Update #437 - accept "+" letter in url parameter

Pull request #437 actually didn't solve the closed issue #435 because I understood the problem wrong. This problem was not about url encoding.

`URLSearchParams` object returns invalid gitignore template id not because the object didn't recognize `+` letters, but because it ACTUALLY recognized them and switched them to SPACES(` `).

To solve the problem correctly, this commit replaces spaces to `+` letters so the filter is able to find proper ids from template set.

* Refactor app.js

This commit changes `replace()` function call position to make it more efficient. It doesn't have to be called for each split value in `templates` parameter one by one. No side effects are anticipated. I moved `toLowerCase()` function as well in the same manner, so `map()` eventually became not required anymore.

This also unifies code style in `app.js` by adding semicolons to some lines. It's not a critical issue of course, according to ECMA and javascript ASI. I just did it to make it look more unified, as the majority lines of original source code already had semicolons.